# updateAPI/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API- for recipes site
  description: This is a simple API
  contact:
    email: adiashk@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.2
servers:
- url: https://virtserver.swaggerhub.com/adiashk/RecipeSystem/1.0.2
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /profile/getIsWatchedRecipe:
    get:
      tags:
      - Profile
      summary: if the user logged he gets info if he whatched the recipe by the id
        of the recipe
      description: if the user logged he gets info if he whatched the recipe by the
        id of the recipe
      parameters:
      - name: recipe_id
        in: query
        description: recipe id checking
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: info if the user whatched the recipe
          content:
            application/json:
              schema:
                type: string
                example: "true"
        "400":
          description: incorrect input
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/get3WatchedRecipes:
    get:
      tags:
      - Profile
      summary: if the user logged he gets 3 recipes, the last thet he watched
      description: if the user logged he gets 3 recipes, the last thet he watched
      responses:
        "200":
          description: 3 last recipes the user whtched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getAllWatchedRecipes:
    get:
      tags:
      - Profile
      summary: if the user logged he gets all the recipes he watched
      description: if the user logged he gets all the recipes he watched
      responses:
        "200":
          description: recipes the user whtched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/addToWatchedRecipes:
    post:
      tags:
      - Profile
      summary: add recipe to the whtched recipes of the user
      description: add recipe to the whtched recipes of the user
      parameters:
      - name: recipe_id
        in: query
        description: add recipe by the id
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: sucessfully add recipes to user whtched
        "400":
          description: incorrect input
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getIsFavoriteRecipe:
    get:
      tags:
      - Profile
      summary: if the user logged he gets info if he saved the recipe by the name
        of the recipe
      description: if the user logged he gets info if he saves the recipe by the name
        of the recipe
      parameters:
      - name: recipe_id
        in: query
        description: recipe id checking
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: info if the user saves the recipe
          content:
            application/json:
              schema:
                type: string
                example: "true"
        "400":
          description: incorrect input
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getFavoritesRecipes:
    get:
      tags:
      - Profile
      summary: get the favorites recipes of the user
      responses:
        "200":
          description: returns all the recipes that the user mark as favorites in
            preview desplay
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/addFavoritesRecipes:
    post:
      tags:
      - Profile
      summary: add to the favorite recipes of the user
      description: Add new recipe to the favorites recipes list of the user
      parameters:
      - name: recipe_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: User added a favorite recipe succesfully
        "400":
          description: incorrect input
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/deleteFavoritesRecipes:
    post:
      tags:
      - Profile
      summary: delete from the favorite recipes of the user
      description: delete recipe from the favorites recipes list of the user
      parameters:
      - name: recipe_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Recipe removed from the favorite recipes list of the user
        "400":
          description: incorrect input
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getMyRecipes:
    get:
      tags:
      - Profile
      summary: Get the user's personal recipes
      responses:
        "200":
          description: returns all the personal recipes of the user in preview desplay
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getMyFullRecipe:
    get:
      tags:
      - Profile
      summary: Get the user's personal recipe by recipe_id
      parameters:
      - name: recipe_name
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: returns the personal recipe of this recipe_id in full desplay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/recipe'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getFullFamilyRecipe:
    get:
      tags:
      - Profile
      summary: Get the user's family recipe by recipe_id in full display
      responses:
        "200":
          description: return family recipe of the user full desplay
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/familyRecipe'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
  /profile/getFamilyRecipes:
    get:
      tags:
      - Profile
      summary: Get the user's family preview recipe
      responses:
        "200":
          description: returns all the family recipes of the user in preview desplay
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/recipePreview'
        "401":
          description: unauthorized - user did not loged in
      security:
      - cookieAuth: []
components:
  schemas:
    recipePreview:
      type: object
      properties:
        id:
          type: integer
          example: 52158
        image:
          type: string
          example: cake.jpg
        name:
          type: string
          example: salad
        preperationTime:
          type: integer
          example: 20
        popularity:
          type: integer
          example: 4
        vegetarian:
          type: boolean
          example: false
        vegan:
          type: boolean
          example: false
        gluten free:
          type: boolean
          example: false
...
    userlogin:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          pattern: '[A-Za-z]*'
          uniqueItems: true
          type: string
          example: adi
        password:
          maxLength: 10
          minLength: 5
          pattern: '[A-Za-z]*[0-9]+[A-Za-z]*[!@#\$%\^\&]+[A-Za-z]*[0-9]*'
          type: string
          format: password
          example: 1a2b!
    recipe:
      type: object
      properties:
        image:
          type: string
          example: cake.jpg
        name:
          type: string
          example: salad
        preperationTime:
          type: integer
          example: 20
        popularity:
          type: integer
          example: 4
        vegetarian:
          type: boolean
          example: false
        vegan:
          type: boolean
          example: false
        gluten free:
          type: boolean
          example: false
        ingredients:
          type: array
          format: string
          example:
          - ingredient: flour
            amount: 2
          - ingredient: oil
            amount: 1.5
          items:
            $ref: '#/components/schemas/ingredient'
        instructions:
          type: array
          items:
            type: string
            example: break the eggs
        servings:
          type: integer
          example: 20
        preperation:
          $ref: '#/components/schemas/userlogin'
    ingredient:
      type: object
      properties:
        ingredient:
          type: string
          example: eggs
        amount:
          type: number
          format: double
          example: 5
    familyRecipe:
      type: object
      properties:
        user_id:
          $ref: '#/components/schemas/user/properties/user_id'
        recipeDetails:
          $ref: '#/components/schemas/recipe'
        whoPrepare:
          type: string
          example: grandmother
        whenPrepare:
          type: string
          example: in holidays
        images:
          type: array
          items:
            type: string
            example: cake.jpg
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie